wifi: mt76: connac: move connac3 definitions in mt76_connac3_mac.h
Connac3 mac definitions are shared between WiFi7 chipsets so move them in
mt76_connac3_mac.h
